+++ src/main/java/com/iexec/worker/compute/post/web2/EncryptionService.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2019-2024 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.iexec.worker.compute.post.web2;
 
 import com.iexec.common.security.EncryptionHelper;
@@ -5,9 +21,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class EncryptionService {
-
     /*
-     *
      * #1: Large file encryption is made with AES
      * #2: AES key is encrypted with RSA key
      *
@@ -17,8 +31,8 @@ public class EncryptionService {
      * after
      * ├── result-0xabc.zip
      * ├── encrypted-result-0xabc
-     * │   ├── aes-key.rsa
-     * │   └── result-0xabc.zip.aes
+     * │   ├── aes-key.rsa
+     * │   └── result-0xabc.zip.aes
      * └── encrypted-result-0xabc.zip (if produceZip)
      *
      * Returns: folder or zip path
@@ -27,29 +41,4 @@ public class EncryptionService {
     public String encryptData(String inDataFilePath, String plainTextRsaPub, boolean produceZip) {
         return EncryptionHelper.encryptData(inDataFilePath, plainTextRsaPub, produceZip);
     }
-
-    /*
-     *
-     * Required: aes-key.rsa file should be found next to encryptedDataFile
-     *
-     * #1: AES key is decrypted with RSA
-     * #2: Data is decrypted with AES key
-     *
-     * before
-     * └── encrypted-result-0xabc.zip
-     * with zip content
-     * ├── aes-key.rsa
-     * └── result-0xabc.zip.aes
-     *
-     * after
-     * ├── encrypted-result-0xabc.zip
-     * └── plain-result-0xabc.zip
-     *
-     * Returns: clear data path (zip here)
-     *
-     * */
-    public String decryptData(String encryptedDataFilePath, String plainTextRsaPriv) {
-        return EncryptionHelper.decryptData(encryptedDataFilePath, plainTextRsaPriv);
-    }
-
 }